ai-service/requirements.txt
fastapi
uvicorn
pdfplumber
pandas
reportlab
python-multipart
requests
google-genai
python-dotenv
groq
